Your Message

+++ packages/KarimQaderi/Zoroaster/src/Http/Controllers/Resource/ResourceDestroyController.php
@@ -12,27 +12,28 @@
 
         public function handle(ResourceRequest $request)
         {
-            foreach(request()->resourceId as $id){
+            foreach(request()->resourceId as $id)
+            {
                 if(method_exists($request->Model() , 'isForceDeleting'))
                     $find = $request->Model()->withTrashed()->where([$request->Model()->getKeyName() => $id])->first();
                 else
                     $find = $request->Model()->where([$request->Model()->getKeyName() => $id])->first();
 
-                $request->authorizeTo($request->Resource()->authorizeToForceDelete($find));
+                if($request->Resource()->authorizedToForceDelete($find))
+                {
+                    $this->ResourceDestroyField($request , $find);
 
-
-                $this->ResourceDestroyField($request , $find);
-
-                if(method_exists($request->Model() , 'isForceDeleting'))
-                    $find->forceDelete();
-                else
-                    $request->Model()->destroy($id);
+                    if(method_exists($request->Model() , 'isForceDeleting'))
+                        $find->forceDelete();
+                    else
+                        $request->Model()->destroy($id);
+                }
 
             }
 
             if(request()->has('redirect'))
                 redirect(request()->redirect)->with([
-                    'success' => 'منبع مورد نظر حذف شد'
+                    'success' => 'منبع مورد نظر حذف شد',
                 ])->send();
 
             return response(['status' => 'ok']);
@@ -43,7 +44,8 @@
             $customResourceController = $request->ResourceFields(function($field){ return true; });
 
             $ResourceError = [];
-            foreach($customResourceController as $field){
+            foreach($customResourceController as $field)
+            {
 
                 $RequestField = new RequestField();
                 $RequestField->request = $request->Request();
@@ -53,7 +55,8 @@
                 $RequestField->MergeResourceFieldsAndRequest = null;
 
                 $ResourceDestroy = (object)$field->ResourceDestroy($RequestField);
-                if(isset($ResourceDestroy->error) && $ResourceDestroy->error !== null){
+                if(isset($ResourceDestroy->error) && $ResourceDestroy->error !== null)
+                {
 
                     if(is_array($ResourceDestroy->error))
                         $ResourceError = array_merge($ResourceError , $ResourceDestroy->error);
