+++ packages/KarimQaderi/Zoroaster/src/Http/Controllers/Resource/ResourceSoftDeletingController.php
@@ -11,19 +11,20 @@
         {
             $cols = null;
             $col = null;
-            foreach(request()->resourceId as $id){
+            foreach(request()->resourceId as $id)
+            {
 
                 $find = $request->Model()->where([$request->Model()->getKeyName() => $id])->first();
 
                 if(is_null($find)) break;
 
-                $request->authorizeTo($request->Resource()->authorizeToDelete($find));
-
-                $find->delete();
+                if($request->Resource()->authorizedToDelete($find))
+                    $find->delete();
 
                 $col = \Zoroaster::ResourceActions($request ,
                     $find ,
-                    $request->Model() , 'Index' , $request->ResourceFields(function($field){
+                    $request->Model() , 'Index' , $request->ResourceFields(function($field)
+                    {
                         if($field !== null && $field->showOnIndex == true)
                             return true;
                         else
@@ -34,9 +35,8 @@
                 $cols [] = [
                     'id' => $find->{$request->Model()->getKeyName()} ,
                     'col' => $col ,
-                    'status' => 'ok'
+                    'status' => 'ok',
                 ];
-
 
 
             }
@@ -44,12 +44,11 @@
 
             if(request()->has('redirect'))
                 redirect(request()->redirect)->with([
-                    'success' => 'منبع مورد نظر حذف شد'
+                    'success' => 'منبع مورد نظر حذف شد',
                 ])->send();
 
             return response($cols);
         }
-
 
 
     }
